% Generated by roxygen2 (4.0.1): do not edit by hand
\name{caculateCooperationIndex}
\alias{caculateCooperationIndex}
\title{Caculating the metabolic competition and complementarity index}
\usage{
caculateCooperationIndex(g, ..., mode = c("competition", "complementarity"),
  threshold = 0)
}
\arguments{
\item{g,igraph}{that represents a metabolic network, see
\code{\link{CnstOrgMtblNet}}}

\item{...,}{a list of metabolic networks or a network append to g}

\item{mode,}{determins which index you want to caculate, competion or
complementarity index}

\item{threshold}{threshold, the cutoff of confidence score to be serve as a
seed set, default is 0.2}
}
\value{
a matrix whose nrow and ncol is equal to the number of species to be
compared, for more see details.
}
\description{
Caculating the metabolic competition complementarity index among all metabolic
networks
}
\details{
Metabolic competition index is defined as the fraction of compounds
in a species seed set of metabolic network that are alse included in its
partner. However, metabolic complementarity index is the fraction of
compounds in one species seed set of metabolic network appearing in the
metabolic network but not in the seed set of its partner.Seed compounds
are associated with a confidence score (1/size of SCC), so this fraction is
calculated as a mormalized weighted sum.

The ith row and jth col elements of the returnd matrix represents the
metabolic competition index or complementarity index of the ith network on the
jth metabolic network.
}

