% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CnstOrgMtblNet}
\alias{CnstOrgMtblNet}
\title{reconstuction of the specific-organism metabolic network}
\usage{
CnstOrgMtblNet(pathway.info)
}
\arguments{
\item{pathway.info,}{df, more details see function "GetOrgPathwayInfo.R"}
}
\value{
igraph object
}
\description{
reconstuction of the specific-organism metabolic network
}

