% Generated by roxygen2 (4.0.1): do not edit by hand
\name{competitionIndex}
\alias{competitionIndex}
\title{Caculating the metabolic competition index}
\usage{
competitionIndex(g1, g2, threshold = 0)
}
\description{
Caculating the metabolic competition index of g1 in the presence of g2
}
\details{
More details abuout the para and value see in complementarityIndex above
}

