% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{\%.\%}
\alias{compose}
\title{Compose multiple functions}
\usage{
compose(...)

f \%.\% g
}
\arguments{
\item{...}{n functions to apply in order from right to left}

\item{f,g}{two functions to compose for the infix form}
}
\description{
In infix and prefix forms.
}
\details{
This function was from hadley wickham's package pryr, for more details see \url{https://github.com/hadley/pryr}
}
\examples{
not_null <- `!` \%.\% is.null
not_null(4)
not_null(NULL)

add1 <- function(x) x + 1
compose(add1,add1)(8)
}
\author{
Hadley wickham
}

