% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getGsMN-methods}
\alias{getGsMN}
\alias{getGsMN-methods}
\title{The genome scale metabolic network}
\usage{
getGsMN(object, field)
}
\arguments{
\item{obj,}{\code{seedset} class}
}
\value{
a igraph
}
\description{
T he genome scale metabolic network (GsMN) whose seed set is caculated.
}
\seealso{
\code{\link{seedset-class}}
}

