% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getOrgMetabolicData.R
\name{getOrgMetabolicData}
\alias{getOrgMetabolicData}
\title{Get organism metabolic data from KEGG database}
\usage{
getOrgMetabolicData(org)
}
\arguments{
\item{org,}{characters, the KEGG organism code, e.g. "buc".}
}
\value{
a three length df, consists of  enzyme reaction names, substrates and
  products
}
\description{
This function helps us  to obtain the specific-organism pathway map, prasing
this maps to get metabolic data contains reaction, substrate and product.
}
\details{
Function \code{getOrgMetabolicData} helps us to download metabolic
  data of a given organism from KEGG database with REST-style KEGG API.
  Enzyme reactions take place in this organism (org) and its metabolites
  (substrates and products), that will be used for organism-specific genome
  scale metabolic network reconstruction, can be obtained with this function.
}
\examples{
\dontrun{
metabolic.data <- getOrgMetabolicData("buc")
}
}
\seealso{
\code{\link{getSeedSets}}
}

