% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetOrgPathwayInfo}
\alias{GetOrgPathwayInfo}
\title{Get organism pathway reaction info}
\usage{
GetOrgPathwayInfo(org)
}
\arguments{
\item{org,}{three characters, the kegg organism code, e.g. "buc"}
}
\value{
a three length df, consists of  reaction name, substrates and products
}
\description{
This function helps us  to obtain the specific-organism pathway map, prasing the maps to get reaction information.
}

