% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getSeedSets}
\alias{getSeedSets}
\title{Identify seed compounds of each organism}
\usage{
getSeedSets(g, threshold = 0.2)
}
\arguments{
\item{g,}{an igraph object which represents a given organism-specific metaboliic network}

\item{threshold,}{numeric constant ranges from 0 to 1}
}
\value{
seed set compounds of the given organism-specific metabolic network
}
\description{
Detect a given metabolic network and idendity the seed compounds of each organism
}
\seealso{
\code{\link{KosarajuSCC}}
}

