% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_buc.R
\name{kegg_buc}
\alias{kegg_buc}
\title{Metabolic profiles of KEGG organism Buchnera aphidicola APS (Acyrthosiphon 
pisum) (KEGG organism code: buc)}
\format{A data frame with 418 observations on three variables.

[,1] .attrs.name, character (reaction: R)

[,2] substrate.name, list (substrates: cpd)

[,3] product.name, list (products: cpd)}
\description{
kegg organism buc metabolic information, which consists of enzymatic reactions
and metabolites.
}
\details{
buc metatolic information:

\itemize{
  \item .attrs.name: Enzymatic reactions that organism involved
  \item substrate.name: Substrates of the corresponding reaction.
  \item product.name: Products of the corresponding reaction.
}
}

