% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kegg_ptr.R
\name{kegg_ptr}
\alias{kegg_ptr}
\title{Metabolic profiles of KEGG organism Pan troglodytes (chimpanzee) (KEGG
organism code: ptr)}
\format{A data frame with 1858 observations on three variables.

 [,1] .attrs.name, character (reaction: R)

 [,2] substrate.name, list (substrates: cpd)

 [,3] product.name, list (products: cpd)}
\description{
kegg organism ptr metabolic information, which consists of enzymatic reactions
and metabolites.
}
\details{
ptr metatolic information:

\itemize{ \item .attrs.name: Enzymatic reactions that organism involved \item
substrate.name: Substrates of the corresponding reaction. \item product.name:
Products of the corresponding reaction. }
}

