% Generated by roxygen2 (4.0.1): do not edit by hand
\name{KosarajuSCC}
\alias{KosarajuSCC}
\title{Caculating the strong connected components (SCC) of a network}
\usage{
KosarajuSCC(g)
}
\arguments{
\item{g,}{a igraph object to be caculated}
}
\value{
a list with three components:
\itemize{
  \item membership  numeric vector giving the SCC id to which each vertex
   belongs.
  \item size numeric vector giving the sizes of the SCCs.
  \item no numeric constant, the number of SCCs.
}
}
\description{
This function utilizes Kosaraju's algorithm to caculate the
strong connetected components descomposition of a given network
}
\references{
\emph{AV Aho, JE Hopcroft, JD Ullman: The design and analysis of computer algorithms, 1974}
}

