% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length-methods}
\alias{length}
\alias{length-methods}
\title{the length of the seed set}
\usage{
length(x)
}
\arguments{
\item{x,}{\code{seed-set} class}
}
\value{
a interger
}
\description{
Caculate the number of the seed source components.
}
\seealso{
\code{\link{seedset-class}}
}

