% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{nonseed-methods}
\alias{nonseed}
\alias{nonseed-methods}
\title{Non seed of the network}
\usage{
nonseed(object, field)
}
\arguments{
\item{obj,}{\code{seedset} class}
}
\value{
a vector
}
\description{
Non seed of the network.
}
\seealso{
\code{\link{seedset-class}}
}

