% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reconstructGsMN}
\alias{reconstructGsMN}
\title{Reconstuction of the specific-organism genome-scale metabolic network}
\usage{
reconstructGsMN(metabolic.data)
}
\arguments{
\item{metabolic.data,}{df, more details see function \code{GetOrgPathwayInfo}
and \code{details}}
}
\value{
igraph object
}
\description{
Reconstruction of genome-scale metabolic network (GsMN) whose
nodes represents compounds and whose edges represents reactions.
}
\details{
The input of this function can be of two forms. If organims is
collected in KEGG database, it can be obtained with \code{GetOrgPathwayInfo}.
If not, \code{metabolic.data} could be a  table format file which contains the
KEGG Orthology annotated information on this organism, e.g. we can download
this KO annotation profile in the \url{https://img.jgi.doe.gov} website for
species detected in a human microbime which not contained in KEGG organism
database. Several functions, such as\code{link{read.table}} and
\code{\link{read.delim}} could help us to read KO annotation profile.
}
\examples{
## not run (organism in KEGG)
## metabolic.data <- getOrgMetabolicData("buc", refresh = TRUE)
## g <- reconstructGsMN(metabolic.data)

## species detected in a human microbiome
annodir <- system.file("extdata/koanno.txt",package = "ReverseEcologyR")
metabolic.data <- read.delim(annodir)
g2 <- reconstructGsMN(ko.anno)
}
\seealso{
\code{\link{getOrgMetabolicData}}
}

