\name{RefDbcache}
\alias{RefDbcache}
\docType{data}
\title{
Reference data for global metabolic construction 
}
\description{
The metabolic pathway data contains KOs, substrates and products, as well as a constructed reference global network  
}
\usage{data(RefDbcache)}
\format{
  The format is:
List of 7
  KO, substrate, product, user, date, version, reference network 
                                                                                                                                                                         : 
}
\examples{
data(RefDbcache)
}
