% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RevEcoR.R
\docType{package}
\name{RevEcoR}
\alias{RevEcoR}
\alias{RevEcoR-package}
\title{The RevEcoR package}
\description{
This package implementation the applications of reverse ecology. Reverse ecology refers to the use of genomics to study ecology with no a priori assumptions about the organism(s) under consideration, linking the organism and their environment. Prediction the cooperation among species and hosts.
}

