% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ReverseEcologyR}
\alias{ReverseEcologyR}
\alias{ReverseEcologyR-package}
\title{The ReverseEcologyR package}
\description{
This package implementation the applications of reverse ecology. Reverse ecology refers to the use of genomics to study ecology with no a priori assumptions about the organism(s) under consideration, linking the organism and their environment.
}

