% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{show,seedset-method}
\alias{show}
\alias{show,seedset-method}
\alias{show-methods}
\title{The show generic function}
\usage{
\S4method{show}{seedset}(object)
}
\arguments{
\item{object,}{\code{seed-set} class}
}
\description{
Show a short summary of seedset object
}
\examples{
\dontrun{
show(seed.set)
}
}
\seealso{
\code{\link{seedset-class}}
}

