% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show-methods}
\alias{show}
\alias{show-methods}
\title{The show generic function}
\usage{
\S4method{show}{seedset}(object)
}
\arguments{
\item{obj,}{\code{seedset} class}

\item{object}{Any R object}
}
\description{
Show a short summary of seedset object
}
\seealso{
\code{\link{seedset-class}}
}

