% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculateCooperationIndex.R
\name{competitionIndex}
\alias{competitionIndex}
\title{Calculating the metabolic competition index}
\usage{
competitionIndex(g1, g2, seed.set1, seed.set2, threshold = 0, p = FALSE,
  nperm = 1000)
}
\arguments{
\item{g1,}{igraph object, a metabolic network}

\item{g2,}{igraph object, a metabolic network, the complementary network of g1}

\item{seed.set1,}{seeds slot of a seed-set object, seeds of the metabolic
network g1, more details see \code{\link{seedset-class}}.}

\item{seed.set2,}{seeds slot of a seed-set object, seeds of the metabolic
network g2, more details see \code{\link{seedset-class}}.}

\item{threshold,}{the cutoff of confidence score to be serve as a seed set,
default is 0.2}

\item{p,}{a logical value which determins whether the calculated index is
statistical or biological significant.
default is FALSE}

\item{nperm,}{the number of permuations of metabolic network node labes, which
is used for complementarity index's P value calculating, default is 1000.}
}
\value{
a two length list: competition index: range from 0 to 1, p value
of competition index. Or a single value of competition index while p
is FALSE
}
\description{
Calculating the metabolic competition index of g1 in the presence of g2
}
\seealso{
\code{\link{complementarityIndex}},
\code{\link{calculateCooperationIndex}}
}

