% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{confidencescore-methods}
\alias{confidencescore}
\alias{confidencescore-methods}
\title{Conficence score}
\usage{
confidencescore(object, field)
}
\arguments{
\item{obj,}{\code{seedset} class}
}
\value{
a vector
}
\description{
Caculate confidence score of seed set
}
\seealso{
\code{\link{seedset-class}}
}

