% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{getGsMN-methods}
\alias{getGsMN}
\alias{getGsMN,seedset-method}
\alias{getGsMN-methods}
\title{The genome scale metabolic network}
\usage{
getGsMN(object)

\S4method{getGsMN}{seedset}(object)
}
\arguments{
\item{object,}{\code{seedset} class}
}
\value{
a igraph
}
\description{
T he genome scale metabolic network (GsMN) whose seed set is caculated.
}
\examples{
\dontrun{
getGsMN(seed.set)
}
}
\seealso{
\code{\link{seedset-class}}
}

