% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOrgMetabolicData}
\alias{getOrgMetabolicData}
\title{Get organism metabolic data from KEGG database}
\usage{
getOrgMetabolicData(org, path = Sys.getenv("home"), refresh = FALSE)
}
\arguments{
\item{org,}{three characters, the KEGG organism code, e.g. "buc".}

\item{path,}{character which specify the dir for saving the local metabolic
data, default is your home dir of R. More information see \code{details}.}

\item{refresh,}{logical, whether refresh the exsits metabolic data from KEGG,
detault is FALSE.}
}
\value{
a three length df, consists of  reaction name, substrates and products
}
\description{
This function helps us  to obtain the specific-organism pathway map, prasing
this maps to get metabolic data contains reaction, substrate and product.
}
\details{
\code{getOrgMetabolicData} helps us to obtain metabolic data of a
given organism from KEGG database. However, keeping a downloaded metabolic
data in local is necessary. It not only can help keeping the data accurate
and speed up the process of network reconstruction, but also allows you to
reuse the data for other purposes.

Futher more, the reference data such as the KO reference hierarchy does not
change frequently.Neither does themetabolic data of organisms whose genome is
sequenced very early. However, to those whose genome is undersequenced or just
got completely annotated completely not very long ago, their metabolic data
is updated frequently. In this case, assigning different TTL (Time-to-live) to
different types of data will reduce network overhead and data retrieval
time caused by remotely visiting KEGG frequently. For example, it would be
helpful to set the renewal period of the KO reference hierarchy data as 30
days while that of the organism-specific.

As \code{getOrgMetabolicData}, check for the existence of queried orgaism
metabolic data in the local database. If the data is not there, retrieve it
from KEGG and write into the local database. If not, return the data
immediately from the local database.
}

