% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KosarajuSCC.R
\name{KosarajuSCC}
\alias{KosarajuSCC}
\title{Caculating the strong connected components (SCC) of a network}
\usage{
KosarajuSCC(g)
}
\arguments{
\item{g, }{a igraph object to be caculated}
}
\value{
a list which length is equal to the number of SCCs, each element represents a Scc
}
\description{
This function utilizes Kosaraju's algorithm to caculate the
strong connetected components descomposition of a given network
}
\examples{
\dontrun{
metabolic.data <- getOrgMetabolicData("buc")
## metabolic network reconstruction
net <- reconstructGsMN(metabolic.data)
scc <- KosarajuSCC(net)
} 
}
\references{
\emph{AV Aho, JE Hopcroft, JD Ullman: The design and analysis of computer algorithms, 1974}
}
\seealso{
\code{\link{getSeedSets}}
}

