% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{len-methods}
\alias{len}
\alias{len,seedset-method}
\alias{len-methods}
\title{the length of the seed set}
\usage{
len(object)

\S4method{len}{seedset}(object)
}
\arguments{
\item{object,}{\code{seed-set} class}
}
\value{
an interger
}
\description{
Caculate the number of the seed source components.
}
\seealso{
\code{\link{seedset-class}}
}

