% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length-methods}
\alias{length}
\alias{length-methods}
\title{the length generic function}
\usage{
\S4method{length}{seedset}(x)
}
\arguments{
\item{obj,}{\code{length} class}

\item{x}{an \R object.  For replacement, a vector or factor.}
}
\value{
a interger
}
\description{
Caculate the number of the seed source components.
}
\seealso{
\code{\link{seedset-class}}
}

