% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{nonseed-methods}
\alias{nonseed}
\alias{nonseed,seedset-method}
\alias{nonseed-methods}
\title{Non seed of the network}
\usage{
nonseed(object)

\S4method{nonseed}{seedset}(object)
}
\arguments{
\item{object,}{\code{seedset} class}
}
\value{
a vector
}
\description{
Non seed of the network.
}
\seealso{
\code{\link{seedset-class}}
}

