% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{seedSize-methods}
\alias{seedSize}
\alias{seedSize-methods}
\title{Size of the each seed source component}
\usage{
seedSize(object)
}
\arguments{
\item{obj,}{\code{seedset} class}
}
\value{
a vector represents size of each source seed componet of network
}
\description{
Caculate the size of each seed source component.
}
\seealso{
\code{\link{seedset-class}}
}

